Move addNote client method to submittedApplication client
Seeing as this is an operation on a submitted application, this is a more suitable location.
